iniciando rotas e ajustando nomes de pastas
