+++ src/app/components/ConnectedUsersCard.tsx
@@ -1,11 +1,28 @@
 import { GiPerson } from "react-icons/gi";
+import { useEffect, useState } from "react";
 
 const ConnectedUsersCard = () => {
-  const users = ["Usuário 2", "Usuário 3", "Usuário 4"]; 
+  const [users, setUsers] = useState([]);
+
+  useEffect(() => {
+    async function fetchData() {
+      try {
+        const response = await fetch("/api/getUsers/routes.tsx");
+        if (!response.ok) {
+          throw new Error(`HTTP error! status: ${response.status}`);
+        }
+        const data = await response.json();
+        setUsers(data);
+      } catch (error) {
+        console.error(error);
+      }
+    }
+
+    fetchData();
+  }, []);
 
   return (
     <div className="bg-secondaryColor border border-primaryColor rounded-3xl p-4 shadow-md max-w-sm mx-auto w-full">
-  
       <h2 className="text-teal-600 text-2xl font-bold mb-4">
         Usuários Conectados
       </h2>
