+++ src/app/components/MobileFooter.tsx
