+++ src/app/components/MobileProfileBar.tsx
