+++ src/index.js
@@ -2,27 +2,25 @@ import React from "react";
 import ReactDOM from "react-dom/client";
 import "./index.css";
 
-
 import { createBrowserRouter, RouterProvider } from "react-router-dom";
 
-import HomePage from "./routes/HomePage";
-import ProfilePage from "./routes/ProfilePage"
+import HomePage from "./app/pages/HomePage";
+import ProfilePage from "./app/pages/ProfilePage";
 
 const router = createBrowserRouter([
   {
     path: "/",
-    element: <HomePage/>
+    element: <HomePage />,
   },
   {
     path: "Profile",
-    element: <ProfilePage/>
+    element: <ProfilePage />,
   },
-])
-
+]);
 
 const root = ReactDOM.createRoot(document.getElementById("root"));
 root.render(
   <React.StrictMode>
-    <RouterProvider router={router}/>
+    <RouterProvider router={router} />
   </React.StrictMode>
 );
